Move ShopReviews component to /components.
